move DragHandle into a separate component
